refactoring sarek models
