+++ ngi_pipeline/engines/sarek/models/sarek.py
@@ -1,18 +1,19 @@
 import csv
 import os
-import re
-from string import Template
 
 from ngi_pipeline.engines.sarek.database import CharonConnector, TrackingConnector
-from ngi_pipeline.engines.sarek.exceptions import BestPracticeAnalysisNotRecognized, ReferenceGenomeNotRecognized, \
-    SampleNotValidForAnalysisError
-from ngi_pipeline.engines.sarek.parsers import QualiMapParser, PicardMarkDuplicatesParser, ParserIntegrator
+from ngi_pipeline.engines.sarek.exceptions import BestPracticeAnalysisNotRecognized, SampleNotValidForAnalysisError
+from ngi_pipeline.engines.sarek.models.resources import ReferenceGenome
+from ngi_pipeline.engines.sarek.models.sample import SarekAnalysisSample
+from ngi_pipeline.engines.sarek.models.workflow import SarekPreprocessingStep, SarekGermlineVCStep, SarekAnnotateStep, \
+    SarekMultiQCStep
+from ngi_pipeline.engines.sarek.parsers import ParserIntegrator
 from ngi_pipeline.engines.sarek.process import ProcessConnector, ProcessRunning, ProcessExitStatusSuccessful, \
     ProcessExitStatusFailed
-from ngi_pipeline.utils.filesystem import safe_makedir, is_index_file
+from ngi_pipeline.utils.filesystem import safe_makedir
 
 
-class SarekAnalysis(object):
+class SarekAnalysis:
     """
     Base class for the SarekAnalysis engine. This class contains the necessary methods for configuring and launching
     an analysis with the Sarek engine. However, some methods are not implemented (they are "abstract") and are
@@ -503,315 +504,4 @@ class SarekGermlineAnalysis(SarekAnalysis):
         return {
             metric: results_parser.query_parsers("get_{}".format(metric)).pop()
             for metric in ["percent_duplication", "autosomal_coverage", "total_reads"]}
-
-
-class SarekAnalysisSample:
-
-    def __init__(self, project_object, sample_object, sarek_analysis, restart_options=None):
-        self.sarek_analysis = sarek_analysis
-        self.sampleid = sample_object.name
-        self.projectid = project_object.project_id
-        self.project_base_path = project_object.base_path
-        self.restart_options = {
-            key: False for key in ["restart_failed_jobs", "restart_finished_jobs", "restart_running_jobs"]}
-        self.restart_options.update(restart_options or {})
-        self.sample_ngi_object = sample_object
-
-    def sample_data_path(self):
-        return self.sarek_analysis.sample_data_path(self.project_base_path, self.projectid, self.sampleid)
-
-    def sample_analysis_path(self):
-        return self.sarek_analysis.sample_analysis_path(self.project_base_path, self.projectid, self.sampleid)
-
-    def sample_seqrun_path(self, libprepid, seqrunid):
-        return os.path.join(self.sample_data_path(), libprepid, seqrunid)
-
-    def sample_analysis_exit_code_path(self):
-        return self.sarek_analysis.sample_analysis_exit_code_path(self.project_base_path, self.projectid, self.sampleid)
-
-    def sample_analysis_tsv_file(self):
-        return self.sarek_analysis.sample_analysis_tsv_file(self.project_base_path, self.projectid, self.sampleid)
-
-    def _get_sample_librep(self, libprepid):
-        return list(filter(lambda lp: lp.name == libprepid, self.sample_ngi_object)).pop()
-
-    def sample_libprep_ids(self):
-        return list(map(lambda x: x.name, self.sample_ngi_object))
-
-    def libprep_seqrun_ids(self, libprepid):
-        return list(map(lambda x: x.name, self._get_sample_librep(libprepid)))
-
-    def libpreps_to_analyze(self):
-        return filter(
-            lambda lp: self.sarek_analysis.libprep_should_be_started(
-                self.projectid, self.sampleid, lp.name),
-            self.sample_ngi_object)
-
-    def seqruns_to_analyze(self, libprep):
-        return filter(
-            lambda sr: self.sarek_analysis.seqrun_should_be_started(
-                self.projectid, self.sampleid, libprep.name, sr.name, self.restart_options),
-            libprep)
-
-    def runid_and_fastq_files_for_sample(self):
-        for libprep in self.libpreps_to_analyze():
-            for libprep_fastq in self._get_runid_and_fastq_files_for_libprep(libprep):
-                yield libprep_fastq
-
-    def _get_runid_and_fastq_files_for_libprep(self, libprep):
-        for seqrun in self.seqruns_to_analyze(libprep):
-            for seqrun_fastq in self._get_runid_and_fastq_files_for_seqrun(libprep.name, seqrun):
-                yield seqrun_fastq
-
-    def _get_runid_and_fastq_files_for_seqrun(self, libprepid, seqrun):
-        # the runfolder is represented with a Runfolder object
-        runfolder = Runfolder(self.sample_seqrun_path(libprepid, seqrun.name))
-        # iterate over the fastq file pairs belonging to the seqrun, filter out files with index reads
-        sample_fastq_objects = map(
-            lambda f: SampleFastq(os.path.join(runfolder.path, f)),
-            filter(lambda fq: not is_index_file(fq), seqrun.fastq_files))
-        for sample_fastq_file_pair in SarekGermlineAnalysis.sample_fastq_file_pair(sample_fastq_objects):
-            runid = "{}.{}.{}".format(
-                runfolder.flowcell_id, sample_fastq_file_pair[0].lane_number, sample_fastq_file_pair[0].sample_number)
-            yield [runid] + map(lambda x: x.path, sample_fastq_file_pair)
-
-
-class SampleFastq(object):
-    """
-    SampleFastq represents a fastq file and takes care of identifying properties such as sample name, lane, read etc.
-    by splitting the path based on an expected regexp.
-    """
-    # the regexp used for splitting the file name
-    FASTQ_REGEXP = r'^(.+)_([^_]+)_L00(\d)_R(\d)[^\.]*(\..*)$'
-
-    def __init__(self, path):
-        """
-        Creates a SampleFastq object representing a fastq file.
-
-        :param path: the path to the fastq file, where the file name will be used to identify properties
-        """
-        self.path = path
-        self.dirname = os.path.dirname(self.path)
-        self.filename = os.path.basename(self.path)
-        self.sample_name, \
-            self.sample_number, \
-            self.lane_number, \
-            self.read_number, \
-            self.file_extension = self.split_filename(self.filename)
-
-    def split_filename(self, filename):
-        """
-        Split the filename according to the regexp and return the identified groups. If the regexp was not matched,
-        None will be returned for all properties.
-
-        :param filename: the filename to split by the regexp
-        :return: a list of the values captured by the regexp or a list of None values if the regexp did not capture
-        anything
-        """
-        match = re.match(self.FASTQ_REGEXP, filename)
-        return match.groups() if match is not None else [None for i in range(5)]
-
-
-class Runfolder(object):
-    """
-    Runfolder represents a runfolder and identifies properties such as run date, instrument id etc. based on the
-    name element of the path and a regexp used to identify them.
-    """
-    # the regexp used to split the runfolder name
-    RUNFOLDER_REGEXP = r'^(\d{6})_([^_]+)_(\d+)_([AB])(\w+)$'
-
-    def __init__(self, path):
-        """
-        Creates a Runfolder object representing a runfolder directory.
-
-        :param path: the path to the runfolder where the name will be used to identify properties
-        """
-        self.path = path
-        self.dirname = os.path.dirname(self.path)
-        self.runfolder_name = os.path.basename(self.path)
-        self.run_date, \
-            self.instrument_id, \
-            self.run_number, \
-            self.flowcell_position, \
-            self.flowcell_id = self.split_runfolder_name(self.runfolder_name)
-
-    def split_runfolder_name(self, runfolder_name):
-        """
-        Split the runfolder name according to the regexp and return the identified groups. If the regexp was not
-        matched, None will be returned for all properties.
-
-        :param runfolder_name: the runfolder name to split by the regexp
-        :return: a list of the values captured by the regexp or a list of None values if the regexp did not capture
-        anything
-        """
-        match = re.match(self.RUNFOLDER_REGEXP, runfolder_name)
-        return match.groups() if match is not None else [None for i in range(5)]
-
-
-class ReferenceGenome(object):
-    """
-    The ReferenceGenome class represents the reference genome used by Sarek for the analysis. It has a factory method
-    for returning the correct instance from a string representation.
-    """
-    NAME = None
-
-    def __repr__(self):
-        return self.NAME
-
-    @staticmethod
-    def get_instance(genome_name):
-        """
-        Factory method to get a ReferenceGenome instance representing the genome from a string.
-
-        :raises: ReferenceGenomeNotRecognized if the genome name was not recognized
-        :param genome_name: the name of the reference as a string, e.g. "GRCh37" or "GRCh38"
-        :return: a ReferenceGenome instance
-        """
-        try:
-            return filter(
-                lambda ref: str(ref).lower() == genome_name.lower(),
-                (GRCh37(), GRCh38())
-            )[0]
-        except IndexError:
-            raise ReferenceGenomeNotRecognized(genome_name)
-
-
-class GRCh37(ReferenceGenome):
-    """Class representing the GRCh37 reference genome"""
-    NAME = "GRCh37"
-
-
-class GRCh38(ReferenceGenome):
-    """Class representing the GRCh38 reference genome"""
-    NAME = "GRCh38"
-
-
-class SarekWorkflowStep(object):
-    """
-    The SarekWorkflowStep class represents an analysis step in the Sarek workflow. Primarily, it provides a method for
-    creating the step-specific command line.
-    """
-
-    available_tools = []
-
-    def __init__(
-            self,
-            path_to_nextflow,
-            path_to_sarek,
-            **sarek_args):
-        """
-        Create a SarekWorkflowStep instance according to the passed parameters.
-
-        :param path_to_nextflow: path to the nextflow executable
-        :param path_to_sarek: path to the main Sarek folder
-        :param sarek_args: additional Sarek parameters to be included on the command line
-        """
-        self.nf_path = path_to_nextflow
-        self.sarek_path = path_to_sarek
-        # create a dict with parameters based on the passed key=value arguments
-        self.sarek_args = {k: v for k, v in sarek_args.items() if k not in ["nf_path", "sarek_path"]}
-        # add/filter a tools parameter against the valid tools for the workflow step
-        self.sarek_args["tools"] = self.valid_tools(self.sarek_args.get("tools", []))
-        # expand any parameters passed as list items into a ","-separated string
-        self.sarek_args = {k: v if type(v) is not list else ",".join(v) for k, v in self.sarek_args.items()}
-
-    def _append_argument(self, base_string, name, hyphen="--"):
-        """
-        Append an argument with a placeholder for the value to the supplied string in a format suitable for the
-        string.Template constructor. If no value exists for the argument name among this workflow step's config
-        parameters, the supplied string is returned untouched.
-
-        Example: step._append_argument("echo", "hello", "") should return "echo hello ${hello}", provided the step
-        instance has a "hello" key in the step.sarek_args dict.
-
-        :param base_string: the string to append an argument to
-        :param name: the argument name to add a placeholder for
-        :param hyphen: the hyphen style to prefix the argument name with (default "--")
-        :return: the supplied string with an appended argument name and placeholder
-        """
-        # NOTE: a numeric value of 0 will be excluded (as will a boolean value of False)!
-        if not self.sarek_args.get(name):
-            return base_string
-        return "{0} {2}{1} ${{{1}}}".format(base_string, name, hyphen)
-
-    def command_line(self):
-        """
-        Generate the command line for launching this analysis workflow step. The command line will be built using the
-        Sarek arguments passed to the step's constructor and returned as a string.
-
-        :return: the command line for the workflow step as a string
-        """
-        single_hyphen_args = ["config", "profile"]
-        template_string = "${nf_path} run ${sarek_step_path}"
-        for argument_name in single_hyphen_args:
-            template_string = self._append_argument(template_string, argument_name, hyphen="-")
-        for argument_name in filter(lambda n: n not in single_hyphen_args, self.sarek_args.keys()):
-            template_string = self._append_argument(template_string, argument_name, hyphen="--")
-        command_line = Template(template_string).substitute(
-            nf_path=self.nf_path,
-            sarek_step_path=os.path.join(self.sarek_path, self.sarek_step()),
-            **self.sarek_args)
-        return command_line
-
-    def sarek_step(self):
-        raise NotImplementedError("The Sarek workflow step definition for {} has not been defined".format(type(self)))
-
-    def valid_tools(self, tools):
-        """
-        Filter a list of tools against the list of available tools for the analysis step.
-
-        :param tools: a list of tool names
-        :return: a list of tool names valid for this workflow step
-        """
-        return list(filter(lambda t: t in self.available_tools, tools))
-
-    @classmethod
-    def report_files(cls, analysis_sample):
-        return []
-
-
-class SarekPreprocessingStep(SarekWorkflowStep):
-
-    def sarek_step(self):
-        return "main.nf"
-
-    @classmethod
-    def report_files(cls, analysis_sample):
-        report_dir = os.path.join(analysis_sample.sample_analysis_path(), "Reports")
-        return [
-            [
-                QualiMapParser,
-                os.path.join(report_dir, "bamQC", analysis_sample.sampleid, "genome_results.txt")],
-            [
-                PicardMarkDuplicatesParser,
-                os.path.join(report_dir, "MarkDuplicates", "{}.bam.metrics".format(analysis_sample.sampleid))]]
-
-
-class SarekGermlineVCStep(SarekWorkflowStep):
-
-    available_tools = [
-        "haplotypecaller",
-        "strelka",
-        "manta"
-    ]
-
-    def sarek_step(self):
-        return "germlineVC.nf"
-
-
-class SarekAnnotateStep(SarekWorkflowStep):
-
-    available_tools = [
-        "snpeff",
-        "vep"
-    ]
-
-    def sarek_step(self):
-        return "annotate.nf"
-
-
-class SarekMultiQCStep(SarekWorkflowStep):
-
-    def sarek_step(self):
-        return "runMultiQC.nf"
